Refactor : Service Interface 추가
