+++ src/main/java/com/hy/demo/Domain/Mail/Service/MailServiceImpl.java
@@ -15,7 +15,7 @@ import javax.mail.internet.MimeMessage;
 
 @Service("async")
 @RequiredArgsConstructor
-public class MailService {
+public class MailServiceImpl {
 
 
     private final JavaMailSender mailSender;
